refactor to separate encryptedIdb and simple Idb storage
